scripts/requirements.txt
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
slack-sdk>=3.0.0
requests>=2.31.0
